Trying a different way to find cantons
Using only university names as a parametrer for geolocators isn't
enough, because we get about half of the results.
A better idea would be using university name, and if there is no result,
using the institution name as a second chance.
We will then take both universities and institution names into account
in order to do our research:
1) Create a key-value (name-canton) dictionary for universities and
institutions :
['University', 'Canton'] and ['Institution', 'Canton']
2) Go trough the dataframe:
- Check if the university name exists in our index. If not, geolocate
the address
- If several addresses are returned, check one that is in Switzerland
(eg: University of Geneva uses to be found is US !)
- If the university address is not relevant, try to find it with
institution name, the same way as above
- Extract the canton of the address (if it found an address and if it's
in Switzerland). If no canton, let's say 'None' canton
- Add the canton name to a the dictionary (or add something like 'None'
if no canton has been found), so next university or institution that has
been already found won't have to be geolocated again
- Add the canton to a table
3) Add the canton table to the above dataframe in a way that they match
with the universities or institutions